Added calls to Ecos that adds the newly created facilities to their respective case. Also changed the handling of the two newest casetypes. Now we are searching for an existing facility and adding that facility to the case
