+++ src/integration-test/resources/EcosCreateCaseIT/__files/common/mappings/ecos_AddFacilityToCase.json
@@ -3,7 +3,7 @@
 		"method": "POST",
 		"bodyPatterns": [
 			{
-				"contains": "AddPartyToCase"
+				"contains": "AddFacilityToCase"
 			}
 		],
 		"url": "/Ecos2PublicServices/MinutMiljoService.svc"
@@ -20,8 +20,8 @@
 			"Content-Type": "text/xml; charset=utf-8",
 			"X-Powered-By": "ASP.NET"
 		},
-		"body": "<s:Envelope xmlns:s=\"http://schemas.xmlsoap.org/soap/envelope/\"><s:Body><AddPartyToCaseResponse  xmlns=\"urn:Ecos.API.MinutMiljo.Service.V1\"/></s:Body></s:Envelope>",
+		"body": "<s:Envelope xmlns:s=\"http://schemas.xmlsoap.org/soap/envelope/\"><s:Body><AddFacilityToCaseResponse  xmlns=\"urn:Ecos.API.MinutMiljo.Service.V1\"/></s:Body></s:Envelope>",
 		"status": 200
 	},
-	"name": "ecos2publicservices_minutmiljoservicesvc_AddFacilityToCase"
+	"name": "ecos2publicservices_minutmiljoservicesvc_AddPartyToCase"
 }
