+++ src/integration-test/resources/EcosCreateCaseIT/__files/test11_createCase_11/mappings/ecos_SearchFacility.json
@@ -3,7 +3,7 @@
 		"method": "POST",
 		"bodyPatterns": [
 			{
-				"contains": "AddPartyToCase"
+				"contains": "SearchFacility"
 			}
 		],
 		"url": "/Ecos2PublicServices/MinutMiljoService.svc"
@@ -20,8 +20,8 @@
 			"Content-Type": "text/xml; charset=utf-8",
 			"X-Powered-By": "ASP.NET"
 		},
-		"body": "<s:Envelope xmlns:s=\"http://schemas.xmlsoap.org/soap/envelope/\"><s:Body><AddPartyToCaseResponse  xmlns=\"urn:Ecos.API.MinutMiljo.Service.V1\"/></s:Body></s:Envelope>",
+		"bodyFileName": "test11_createCase_11/response/EcosSearchFacility.xml",
 		"status": 200
 	},
-	"name": "ecos2publicservices_minutmiljoservicesvc_AddFacilityToCase"
+	"name": "ecos2publicservices_minutmiljoservicesvc_SearchFacility"
 }
