+++ src/integration-test/resources/EcosCreateCaseIT/__files/test11_createCase_11/response/EcosSearchFacility.xml
@@ -1,0 +1,20 @@
+<s:Envelope xmlns:s="http://schemas.xmlsoap.org/soap/envelope/">
+	<s:Body>
+		<SearchFacilityResponse xmlns="urn:Ecos.API.MinutMiljo.Service.V1">
+			<SearchFacilityResult xmlns:a="urn:Ecos.API.MinutMiljo.DataContracts.V1" xmlns:i="http://www.w3.org/2001/XMLSchema-instance">
+				<a:SearchFacilityResultSvcDto i:type="a:SearchFacilityResultSvcDto">
+					<a:EstateDesignation>Test</a:EstateDesignation>
+					<a:FacilityCollectionName>Test</a:FacilityCollectionName>
+					<a:FacilityId>Test</a:FacilityId>
+					<a:FacilityName>e19981ad-34b2-4e14-88f5-133f61ca85aa</a:FacilityName>
+					<a:FacilityStatusName>Test</a:FacilityStatusName>
+					<a:FacilityTypeName>Test</a:FacilityTypeName>
+					<a:Place>Test</a:Place>
+					<a:PostCode>Test</a:PostCode>
+					<a:PostalArea>Test</a:PostalArea>
+					<a:VisitingAddress>Test</a:VisitingAddress>
+				</a:SearchFacilityResultSvcDto>
+			</SearchFacilityResult>
+		</SearchFacilityResponse>
+	</s:Body>
+</s:Envelope>
